requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Database drivers
motor==3.3.2
redis[hiredis]==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HuggingFace
huggingface-hub==0.20.3

# Utilities
python-dateutil==2.8.2

# Rate limiting
slowapi==0.1.9

# Monitoring (optional)
prometheus-fastapi-instrumentator==6.1.0